requirements.txt
# SmartReach Consolidated Requirements
# Python 3.8+ required

# Database
psycopg2-binary>=2.9.9

# Core Data Processing
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0

# Web Scraping & Browser Automation
playwright>=1.40.0
beautifulsoup4>=4.12.2
requests>=2.31.0

# APIs & Trading
yfinance>=0.2.18
alpaca-trade-api>=3.0.0
tweepy>=4.14.0

# SEC Data
sec-edgar-downloader>=5.0.0

# Machine Learning & NLP
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.5.0

# Network Analysis
networkx>=3.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
emoji>=2.8.0

# AI/LLM Integration (if using Anthropic)
anthropic>=0.3.0